+++ server/src/parserClasses.ts
@@ -5,7 +5,9 @@ import { WorkDoneProgressServerReporter } from "vscode-languageserver";
  * A position in a text document expressed as zero-based line and character offset.
  */
 export class Position {
+    /** The line number of the position (zero-based). */
     line: number;
+    /** The character offset of the position (zero-based). */
     character: number;
     constructor(line: number, character: number) {
         this.line = line;
@@ -143,7 +145,10 @@ export enum MatchType {
     localCycleCallName = "localCycleCallName",
     globalCycleCall = "globalCycleCall",
     globalCycleCallName = "globalCycleCallName",
-
+    cycleParameter = "cycleParameter",
+    cycleParameterAssignment = "cycleParameterAssignment",
+    cycleParamList = "cycleParamList",
+    
     controlBlock = "controlBlock",
     gotoBlocknumber = "gotoBlocknumber",
     gotoLabel = "gotoLabel",